update v1.7 with webinterface
