+++ rpi1wire/_pv_1_6_0/locale.yaml
@@ -1,9 +1,8 @@
 # translations for the web interface
 plugin_translations:
     # Translations for the plugin specially for the web interface
-    'Wert':         {'de': '=', 'en': 'Value'}
-    'Aktualisiert': {'de': '=', 'en': 'Updated'}
-    'Pfad': {'de': '=', 'en': 'Directory'}
+    'Wert 2':         {'de': '=', 'en': 'Value 2'}
+    'Wert 4':         {'de': '=', 'en': 'Value 4'}
 
     # Alternative format for translations of longer texts:
     'Hier kommt der Inhalt des Webinterfaces hin.':
